+++ diff-coverage/src/funcTest/kotlin/com/form/plugins/TestingFiles.kt
@@ -1,5 +1,6 @@
 package com.form.plugins
 
+import org.junit.rules.TemporaryFolder
 import java.io.File
 
 inline fun <reified T> getResourceFile(filePath: String): File {
@@ -7,3 +8,17 @@ inline fun <reified T> getResourceFile(filePath: String): File {
             .getResource(filePath)!!.file
             .let(::File)
 }
+
+inline fun <reified T> TemporaryFolder.copyResourceFile(fileFrom: String, fileTo: String): File {
+    return getResourceFile<T>(fileFrom).copyTo(newFile(fileTo), true)
+}
+
+inline fun <reified T> TemporaryFolder.copyResourceDir(dirToCopy: String, destDir: String): String {
+    getResourceFile<T>(dirToCopy).copyRecursively(
+        newFolder(destDir),
+        true
+    )
+    return root.resolve(destDir).toUnixAbsolutePath()
+}
+
+fun File.toUnixAbsolutePath() : String = absolutePath.replace("\\", "/")
