+++ diff-coverage/src/funcTest/resources/git-diff-source-test-files/Class1GitTest.java
