+++ diff-coverage/src/funcTest/resources/single-module-test-project/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'com.form.diff-coverage'
+    id 'java'
+    id 'jacoco'
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation 'junit:junit:4.12'
+}
