+++ diff-coverage/src/funcTest/resources/single-module-test-project/src/main/java/com/java/test/Class1.java
