+++ diff-coverage/src/funcTest/resources/single-module-test-project/src/test/java/com/java/test/Class1Test.java
